requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
Pillow==9.5.0
mysqlclient==2.1.1
redis==4.5.0
celery==5.2.0
python-decouple==3.8
django-filter==23.2
markdown==3.4.3